+++ python/tic-toc-toe/app/minimax.py
@@ -1,81 +1,14 @@
-""" Module contains logic of Tic Toc Toe"""
-import logging
-import os
-import sys
-import tkinter as tk
-from tkinter import messagebox
+"""minimax algoritm"""
 
 
-class Window:
-    """class Window"""
-
-    def __init__(self, height, width) -> None:
-        self.height = height
-        self.width = width
-        self.window = tk.Tk()
-        # status bar
-        self.text_place = tk.StringVar()
-        self.status = tk.Label(
-            self.window,
-            textvariable=self.text_place,
-            fg="green",
-            font="Courier 30",
-        )
-        self.status.place(x=70, y=360)
-
-    def createplace(self):
-        """create main window"""
-        self.window.title("Tic toc toe")
-        self.window.geometry(f"{self.width}x{self.height}")
-
-    def show(self):
-        """show work place"""
-        self.window.mainloop()
-
-
-class Menu:
-    """class Menu"""
-
-    @staticmethod
-    def readlog():
-        """returne logfile"""
-        try:
-            with open("xolog.log", "r") as file:
-                logs = file.read()
-                messagebox.showinfo("История", logs)
-        except FileNotFoundError:
-            messagebox.showinfo("История", "Пока пусто")
-
-    @staticmethod
-    def dellog():
-        """delete logfile"""
-        if os.path.exists("xolog.log"):
-            os.remove("xolog.log")
-            messagebox.showinfo("Info", "Файл 'xolog.log' удален")
-        else:
-            messagebox.showinfo("Info", "Файл 'xolog.log' не найден или не существует ")
-
-    @staticmethod
-    def exit():
-        """exit game"""
-        sys.exit()
-
-
-class Gameprocesse:
-    """class Gameprocesse"""
+class Minimax:
+    """class Minimax"""
 
     def __init__(self) -> None:
         self.place = None
         self.AI = "X"
         self.PLAYER = "O"
         self.oponent = None
-
-    @staticmethod
-    def new_game(place):
-        """update self.place (work place)"""
-        for i, value in enumerate(place):
-            for j, val in enumerate(value):
-                place[i][j]["text"] = " "
 
     @staticmethod
     def check_win(place):
@@ -94,26 +27,7 @@ class Gameprocesse:
                 return place[0][2]["text"]
         return False
 
-    def show_winner(self, place, counter, status):
-        """show winner in mb and refresh status"""
-        if counter > 4:
-            if self.check_win(place):
-                if counter % 2 == 0:
-                    messagebox.showinfo("Игра окончена", "Победил Игрок2")
-                    status = False
-                    return True
-                else:
-                    messagebox.showinfo("Игра окончена", "Победил Игрок1")
-                    status = False
-                    return True
-
-        if counter == 9:
-            messagebox.showinfo("Игра окончена", "Ничья")
-            status = False
-            return True
-
     # MINIMAX
-    #################################
     @staticmethod
     def freecolumn(places):
         """return free spots"""
@@ -178,17 +92,3 @@ class Gameprocesse:
                             move = i
 
         return move
-
-
-# ------------------------------------------------------------------
-class Loging:
-    """class Loging"""
-
-    @staticmethod
-    def log_win(name):
-        """write logfile of winners"""
-        logging.basicConfig(
-            format="%(asctime)s - %(message)s", level=logging.INFO, filename="xolog.log"
-        )
-        logger = logging.getLogger()
-        logger.info(f"{name} победил!")
